Support extending js-yaml Schemas in getSchema().
This adds an argument to the getSchema() function to allow passing in a
list of existing js-yaml Schema objects to produce an extended schema
from.

The actual logic doing this is handled entirely by js-yaml itself, it's
Schema.create() function has a two-argument variant that takes a list of
Schema objects as the first argument and a list of Type objects as the
second argument, and produces a schema consisting of the existing types
defined in the Schema objects passed in plus the list of Type objects
passed in.

By default, this uses the DEFAULT_SAFE_SCHEMA from js-yaml as the Schema
to extend, which produces behavior identical to calling Schema.create()
with just a list of extension types.

By adding this functionality, yaml-import can be used with custom types,
or with the special types provided by js-yaml for Javascript regular
expressions and functions.